+++ control-base/src/main/java/fi/nls/oskari/control/data/MapLayerGroupsHandler.java
@@ -13,7 +13,7 @@ import org.json.JSONObject;
 import org.oskari.log.AuditLog;
 import org.oskari.service.util.ServiceFactory;
 
-import fi.mml.map.mapwindow.service.db.OskariMapLayerGroupService;
+import org.oskari.service.maplayer.OskariMapLayerGroupService;
 import fi.mml.map.mapwindow.service.db.OskariMapLayerGroupServiceIbatisImpl;
 import fi.nls.oskari.annotation.OskariActionRoute;
 import fi.nls.oskari.control.ActionException;
