+++ service-admin/src/main/java/org/oskari/admin/MapLayerGroupsHelper.java
@@ -1,6 +1,6 @@
 package org.oskari.admin;
 
-import fi.mml.map.mapwindow.service.db.OskariMapLayerGroupService;
+import org.oskari.service.maplayer.OskariMapLayerGroupService;
 import fi.nls.oskari.domain.map.MaplayerGroup;
 import fi.nls.oskari.map.layer.group.link.OskariLayerGroupLink;
 import fi.nls.oskari.map.layer.group.link.OskariLayerGroupLinkService;
