+++ service-csw/src/main/java/fi/nls/oskari/control/metadata/MaplayerGroupHandler.java
@@ -1,6 +1,6 @@
 package fi.nls.oskari.control.metadata;
 
-import fi.mml.map.mapwindow.service.db.OskariMapLayerGroupService;
+import org.oskari.service.maplayer.OskariMapLayerGroupService;
 import fi.mml.portti.service.search.SearchCriteria;
 import fi.nls.oskari.domain.map.MaplayerGroup;
 import fi.nls.oskari.service.OskariComponentManager;
